Keep track of FAT overwrite order, don't scan
Instead of looking on-flash each time we update the FAT, keep an ordered
list in memory of the order in which we should be writing to the FAT copies.

This also adds full validation of FAT sectors before putting them in the
list, so invalid (i.e. power lost inmiddle of erase or write) are
overwritten first.